Add Solution to StatusServiceImpl, RoleServiceImpl, ProjectServiceImpl and IssueServiceImpl

+++ src/main/java/ua/duikt/learning/java/pro/spring/repositories/UserRoleRepo.java
@@ -8,5 +8,5 @@ import ua.duikt.learning.java.pro.spring.entity.UserRole;
  * email mykyta.sirobaba@gmail.com
  */
 public interface UserRoleRepo extends JpaRepository<UserRole, UserRole> {
-    long deleteByUserIdAndRoleId(Integer userId, Integer roleId);
+    long deleteByUserIdAndRoleId(Long userId, Long roleId);
 }
